backend/pyproject.toml
[project]
name = "crypto-monitor-backend"
version = "0.1.0"
description = "专业的加密货币异动监控系统 - 后端服务"
authors = [{ name = "Your Name", email = "your@email.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-openai>=1.0.1",
    "watchdog>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=24.1.0",
    "matplotlib>=3.10.8",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.24.0",
    "black>=23.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "modules"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
